3.8.6
filter index allows fast search on fields with exact values
New statistical functions for Relation: BigramStat(), Levenshtein(),
TrigramSimilarity(), TrigramWeightedSimilarity(),
TrigramCosineSimilarity()
Faster link parser for internal links
Fixed table genration in style parser (multiple lines)
Fixed errors on incomplete bloom and monogram indexes
Optional $swWikiTextPre=false to opt out pre tags in wikitext (line
starting with space)

+++ inc/fulltext.php
@@ -134,7 +134,7 @@ function swQueryFulltext($query)
     		$tp = new swTuple($d);
 			$r->tuples[$tp->hash()] = $tp;
 		}
-		else
+		elseif ($db->indexedbitmap->getbit($rev))
 		{
 			$q = "DELETE FROM pages WHERE revision = $rev";
 			if (!$swFulltextIndex->exec($q))
